gvt/requirements.txt
deepspeed==0.9.2
numpy==1.23.4
pandas==2.0.1
Pillow==9.5.0
pyarrow==11.0.0
pycocotools==2.0.6
pytorch_lightning==1.8.5
sacred==0.8.4
timm==0.4.12
torch==2.0.0+cu118
torchvision==0.15.1+cu118
tqdm==4.65.0
transformers==4.28.0.dev0
transformers.egg==info
xformers.egg==info
einops
einops_exts
xformers
